Added SignalDispatcher
